Add environment support to puppet.run_agent and puppet.show_changes
The environment to use can be defined using parameter "environment". This allows
testing what would happen if changes made to modules in testing environment were
merged to production. That said, if Puppet encounters any errors during the
agent run, it will switch back to the environment defined in puppet.conf, which
is usually "production".

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>